Switch inputs, routers and triggers to use a dynamic type registry
